+++ CustomRunOptions/Features/NewRunOptionsButton.cs
@@ -35,7 +35,7 @@ internal sealed class NewRunOptionsButton : IRegisterable
 			.. ModEntry.Instance.CustomRunOptions.Select(o => o.MakeCustomRunSettings(options, g, config)),
 			ModEntry.Instance.ModSettingsApi.MakeBackButton(),
 		]);
-		options.subRoute = ModEntry.Instance.ModSettingsApi.MakeModSettingsRoute(setting);
+		options.subRoute = new WrapperRoute(ModEntry.Instance.ModSettingsApi.MakeModSettingsRoute(setting));
 	}
 
 	private static void NewRunOptions_RenderWarning_Prefix_First()
@@ -110,6 +110,25 @@ internal sealed class NewRunOptionsButton : IRegisterable
 			xPosition += (int)textRect.w;
 			
 			BigNumbers.Render(elements.Count - elementsFitting, xPosition, yPosition - 4, Colors.textMain);
+		}
+	}
+
+	private sealed class WrapperRoute(Route route) : Route
+	{
+		public override MusicState? GetMusic(G g)
+			=> Music.@void;
+
+		public override void Render(G g)
+			=> route.Render(g);
+
+		public override bool TryCloseSubRoute(G g, Route r, object? arg)
+		{
+			if (r != route)
+				return base.TryCloseSubRoute(g, r, arg);
+
+			r.OnExit(g.state);
+			g.CloseRoute(this, arg);
+			return true;
 		}
 	}
 }